Add support for Clang Static Analyzer
The Clang Static Analyzer[1] encounters some issues parsing
GoogleTest's macros, mostly where GoogleTest has deliberately
attempted to obfuscate the behavour of some code to work around
compiler warnings. Specifically, GTest defines a AlwaysTrue() function
which in reality always returns true, but who's definition lives in a
seperate compilation unit which prevents compilers (and Clang
analyzer) from knowing that certain paths are always (or never)
executed.

The net effect of this is that Clang analyzer raises a number of false
positives. For example consider code of the form:

    1    Foo* foo = nullptr;
    2    ASSERT_NO_THROW(foo = some_function());
    3    foo->bar;

In practice foo is guaranteed to be non-NULL at line 3, as if the
assignment at line 2 failed then we wouldn't have reached line
3. However the expansion of the ASSERT_NO_THROW macro involves code of
the form:

    if (AlwaysTrue()) {
        try { foo = ... }
    } ...

The analyzer doesn't know that the conditional will /always/ be
executed, so it belives there's a path where foo is not initialized at
line 3 and hence raises a report.

Solve this by making the function inline when running with clang
analyzer, un-hiding the definition and allowing Clang to see that the
conditional path is actually unconditional.

[1]: http://clang-analyzer.llvm.org

Change-Id: I72906e9f1f9c283bc52178730c99bf171a7c966a